Multi-laned side done. Todo: switching lanes and reconstructing car vector.

+++ car.cpp
@@ -3,11 +3,11 @@
 
 #include "car.hpp"
 
-int Car::update(int nearest_barrier, 
-		int nearest_barrier_speed,
-		int update_frequency,
-		int recommended_speed,
-		int safe_distance)
+size_t Car::update(size_t nearest_barrier, 
+		size_t nearest_barrier_speed,
+		size_t update_frequency,
+		size_t recommended_speed,
+		size_t safe_distance)
 {
 	if (nearest_barrier < critical_distance)
 	{
@@ -17,8 +17,8 @@ int Car::update(int nearest_barrier,
 		}
 		return 0;
 	}
-	int passed_distance = 0;
-	int previous_speed = speed;
+	size_t passed_distance = 0;
+	size_t previous_speed = speed;
 	if (nearest_barrier <= safe_distance)
 	{
 		speed = (nearest_barrier / update_frequency) - previous_speed;
@@ -35,12 +35,12 @@ int Car::update(int nearest_barrier,
 	return passed_distance;
 }
 
-int Car::get_speed() const
+size_t Car::get_speed() const
 {
 	return speed;
 }
 
-int Car::get_size() const
+size_t Car::get_size() const
 {
 	return car_size;
 }
